% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytall.R
\name{get_articles}
\alias{get_articles}
\title{Collect NYT Article Search API Data by Term for Select Years}
\usage{
get_articles(term, start, nyt_key, years = NULL)
}
\arguments{
\item{term}{A string defining the search term. Use single apostrophes around
term for exact matches.}

\item{start}{A numeric year (YYYY) to start the search in.}

\item{nyt_key}{A string defining your New York Times API key:
https://developer.nytimes.com/}

\item{years}{(Optional) A numeric number of years to search after the start
year. Default or NULL gets all data from start to current year.}
}
\value{
A dataframe of one article per row with various text/meta columns.
}
\description{
This function returns a dataframe of New York Times article data for a given
term and selected years. Gets up to 2,000 articles in each search year at up
to 75 articles/min. Updates will be printed in the console every 8 seconds.
}
\examples{
# df <- get_articles("conservation", 1960, nyt_key)
# df <- get_articles("'climate change'", 1960, nyt_key)
# df <- get_articles("'climate change'", 1960, nyt_key, 10)
}
